requirements.txt
torch
torchvision
git+https://github.com/openai/CLIP.git
Pillow
numpy
tqdm
scikit-learn
Flask